requirements.txt
numpy==1.21.5
Pillow==9.0.0
streamlit==1.4.0
streamlit_drawable_canvas==0.8.0
tensorflow==2.7.0
~treamlit==1.3.1
